support nullable case, such as {ifUndefined: null}
